refactor(ecs): move dispatcher to system dir
