+++ core/samples/ecs/general/main.cpp
@@ -6,7 +6,7 @@
 #include <cubos/core/ecs/component/null_storage.hpp>
 #include <cubos/core/ecs/component/registry.hpp>
 #include <cubos/core/ecs/component/vec_storage.hpp>
-#include <cubos/core/ecs/dispatcher.hpp>
+#include <cubos/core/ecs/system/dispatcher.hpp>
 #include <cubos/core/ecs/system/system.hpp>
 #include <cubos/core/ecs/world.hpp>
