+++ core/src/cubos/core/ecs/system/dispatcher.cpp
@@ -1,6 +1,6 @@
 #include <algorithm>
 
-#include <cubos/core/ecs/dispatcher.hpp>
+#include <cubos/core/ecs/system/dispatcher.hpp>
 
 using namespace cubos::core::ecs;
